# apps/api/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/projects": {
      "get": {
        "operationId": "ProjectsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque; from a previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "domain",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "slug": {
                            "type": "string",
                            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "domain": {
                            "type": "string",
                            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                          },
                          "role": {
                            "type": "string",
                            "minLength": 1
                          },
                          "period": {
                            "type": "object",
                            "properties": {
                              "from": {
                                "type": "string",
                                "pattern": "^\\d{4}-\\d{2}$"
                              },
                              "to": {
                                "type": "string",
                                "pattern": "^\\d{4}-\\d{2}$"
                              }
                            },
                            "required": [
                              "from"
                            ],
                            "additionalProperties": false
                          },
                          "summary": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 280
                          },
                          "stack": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "minItems": 1
                          },
                          "visibility": {
                            "type": "string",
                            "enum": [
                              "public",
                              "private",
                              "client"
                            ]
                          },
                          "featured": {
                            "type": "boolean"
                          },
                          "order": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "outcome": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "label": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "value": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "placeholder": {
                                  "type": "boolean",
                                  "default": false
                                }
                              },
                              "required": [
                                "label",
                                "value"
                              ],
                              "additionalProperties": false
                            },
                            "default": []
                          },
                          "links": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "label": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "url": {
                                  "type": "string",
                                  "format": "uri"
                                }
                              },
                              "required": [
                                "label",
                                "url"
                              ],
                              "additionalProperties": false
                            },
                            "default": []
                          },
                          "formation": {
                            "type": "string",
                            "enum": [
                              "monolith",
                              "stream",
                              "lattice",
                              "orbit",
                              "scatter",
                              "grid",
                              "ring",
                              "badge"
                            ],
                            "default": "badge"
                          },
                          "placeholder": {
                            "type": "boolean",
                            "default": false
                          }
                        },
                        "required": [
                          "slug",
                          "name",
                          "domain",
                          "role",
                          "period",
                          "summary",
                          "stack",
                          "visibility",
                          "featured",
                          "order"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "data",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "content"
        ]
      }
    },
    "/v1/projects/{slug}": {
      "get": {
        "operationId": "ProjectsController_get",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1
                    },
                    "domain": {
                      "type": "string",
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                    },
                    "role": {
                      "type": "string",
                      "minLength": 1
                    },
                    "period": {
                      "type": "object",
                      "properties": {
                        "from": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}$"
                        },
                        "to": {
                          "type": "string",
                          "pattern": "^\\d{4}-\\d{2}$"
                        }
                      },
                      "required": [
                        "from"
                      ],
                      "additionalProperties": false
                    },
                    "summary": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 280
                    },
                    "stack": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      },
                      "minItems": 1
                    },
                    "visibility": {
                      "type": "string",
                      "enum": [
                        "public",
                        "private",
                        "client"
                      ]
                    },
                    "featured": {
                      "type": "boolean"
                    },
                    "order": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "outcome": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string",
                            "minLength": 1
                          },
                          "value": {
                            "type": "string",
                            "minLength": 1
                          },
                          "placeholder": {
                            "type": "boolean",
                            "default": false
                          }
                        },
                        "required": [
                          "label",
                          "value"
                        ],
                        "additionalProperties": false
                      },
                      "default": []
                    },
                    "links": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string",
                            "minLength": 1
                          },
                          "url": {
                            "type": "string",
                            "format": "uri"
                          }
                        },
                        "required": [
                          "label",
                          "url"
                        ],
                        "additionalProperties": false
                      },
                      "default": []
                    },
                    "formation": {
                      "type": "string",
                      "enum": [
                        "monolith",
                        "stream",
                        "lattice",
                        "orbit",
                        "scatter",
                        "grid",
                        "ring",
                        "badge"
                      ],
                      "default": "badge"
                    },
                    "placeholder": {
                      "type": "boolean",
                      "default": false
                    }
                  },
                  "required": [
                    "slug",
                    "name",
                    "domain",
                    "role",
                    "period",
                    "summary",
                    "stack",
                    "visibility",
                    "featured",
                    "order"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string",
                      "minLength": 1
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "instance",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "content"
        ]
      }
    }
  },
  "info": {
    "title": "Portfolio API",
    "description": "Public read API behind Wieslaw Samushonga's portfolio. Content is seeded from the site's git repository and served with strong ETags and CDN cache headers. Every error is RFC 9457 Problem Details (`application/problem+json`) carrying the request ID from `X-Request-Id`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.example.dev"
    }
  ],
...
}